feat: add Atmosphere, SkyDome and Glow entities
The `Atmosphere` entity is a realistic looking entity that simulates
sun scattering in the atmosphere. It depends on the sun direction.

The `Glow` entity is a very simple ellipsoidal glow that can be used
to simulate globes that emit light (such as stars).

The `SkyDome` entity simulates the sky and sun disk from inside the
atmosphere.