feat(Map): add support for side skirts (#551)
Map skirts are additional vertices on each map tile, forming vertical
walls on each side, as well as the bottom floor. Vertices for skirts
are put after the vertices for the top side, so that code that deal
with those vertices is not affected.

It serves to give a sense of volume to maps, and the skirt's floor
elevation can be configured, so that for example we can represent the
height of a mountain from the altitude of zero.

In the shader, we distinguish between the top side ("surface") and the
rest for rendering. All fragments that do not belong to the surface are
rendered using the background color only (no layer, or effects like
graticule and contour lines). The non-surface vertices have their own
normal, so that it behaves correctly with lighting and shading.

Note: skirts are per-tile, and not only on the sides of the map. Meaning
that it will behave correctly if the map is cut by a clipping plane for
example.

fix #551