perf(TileMesh): force heightmap use when min/max is not precise enough (#515)
Since map tiles inherit their parent's min/max when created (as they don't
have their own data yet), the volume of the new tile is naturally an
elongated rectangle in the vertical axis (since the tile is twice as small
as its parent, but its elevation range is the same).

This can lead to excessive subdivisions when the camera is close to the
tile's volume (which is likely to happen since it's much taller than
the actual geometry).

To mitigate this problem, we now compute the heightmap if the ratio
between the tile's size and its height is greater than a threshold.
This ensures that we have a tighter volume that matches closely with the
surface.

Previously, the heightmap was computed only when querying elevation or
performing raycasting.

fix #515