fix(Instance): honor entity parent when adding/removing (#531)
When Instance.add() is called with an Object3D or an Entity3D, it will
be added to the default scene graph sub-tree only if it does not already
has a parent. Otherwise this parent is honored.

Conversely, Instance.remove() now correctly removes an object that is
not in the default sub-tree (for example if it has been moved by the user).

This enables user to move the entities and objects freely in the scene
graph after and before adding them to the instance.

fix #531