fix(Shape): ensure that Vector3s passed to the Shape are clone()d
It's absolutely vital that all Vector3's in the .points array of the
Shape are different instances, even for points that are identical (i.e
the first and last point of a closed shape), because those points might
be transformed using a matrix. Having two points being the same reference
mean that the matrix will be applied twice (once on the first point, once
on the last), instead of once.