fix(ColorLayer): fix flickering when the source is synchronous (#525)
In the case where a map contains both an elevation layer and a
color layer with a synchronous* source, two different logics conflict:

When the source is not synchronous, then when subdividing a tile, we
have to inherit the parent tile's texture while the source is loading.

When the source is synchronous, **we don't inherit the parent's
texture** since we immediately create the correct texture.

However, when the elevation texture is not loaded for a given tile, we
also **don't start loading the color texture** (because this tile might
be only a temporary tile until the subdivision is finished).

This optimization (introduced with a716a84f2) is designed to avoid
unnecessary HTTP requests for a tile that will be hidden very soon.

So we might end up with a transient map tile that does not have any
color texture because the source was not allowed to load and also we
didn't inherit the parent's texture.

The solution is simply to allow loading color textures of transient
tiles if the source is synchronous, since this operation is very cheap.

*a synchronous `ImageSource`, introduced in 480e7ff6, is able to
immediately return a texture when asked, without returning a promise.
This is the case of `VectorSource`, since the vector data is already
loaded and not HTTP request is required.

fix #525