feat(DrawTool): allow ending drawing on any input (default: double-click) (#541)
The ending condition is now a custom callback. The previous default
ending condition was right-click, but since it clashes with camera
rotation is most settings, it has now been changed for the double-click.

BREAKING CHANGE: The trigger to end a drawing with the `DrawTool` is
now a double-click instead of right click, to allow users rotating the
camera while drawing. This trigger is now configurable.

fix #541