feat(FeatureCollection): support shading through the style API
Until now, whether to use shaded or unshaded materials was left to the
decision of the GeometryConverter: if a polygon was extruded, then we
used shaded materials. However, we also need to use shaded materials for
non 2D polygons (walls, roofs, etc) even without extrusion.

Now, shading has to be excplitely enabled in the feature styles, and
no longer depends on the fact that the polygon is extruded or not.

The FillStyle.shading property allows users to specify if the feature
should be shaded or not.