fix(DrawTool): ensure that the new shape is immediately displayed (#547)
When starting the creation of a shape without moving the mouse (e.g by
a keyboard shortcut), then the new shape would not appear on the screen
until the first mousemove event was received (so that we know where
to put the first temporary point).

Now the DrawTool track the mouse position on the window at any time so
that we can properly display the shape under the mouse, even when no
mousemove event occured after starting the creation.

fix #547