fix(WorkerPool): use a better metric to select the least busy worker
The previous metric would use the `progress` property of `OperationCounter`,
but this value would be the same for all workers (i.e zero) until the
first task is completed. This would lead to overselecting the
first worker, whereas the other "equally busy" workers would remain
almost empty.

Now we use the `OperationCounter.operations` property, which is much
more accurate at representing how busy a worker is.