fix(PointCloudMaterial): stop applying the offset scale from the layer
The offset scale does not apply for point clouds because of the way the
texture is applied. There are no UV coordinates in point clouds. Instead,
we compute the location of the point relative to the extent, which
acts as a texture coordinate.

Applying the offset scale would give incorrect results (scaled textures
that do not match their extent).