fix(PickObjectsAt): add option to provide custom raycaster (#572)
In the case of point clouds, a point is considered as picked when its
distance to the ray is less than `raycaster.params.Points.threshold`,
which is by default 1 world unit (typically 1 meter). In the case of
small point clouds with lots of details, this will lead to many incorrect
picked points.

However, since there is no reliable way to compute a correct threshold
in all cases, we now provide the option for the user to pass their own
raycaster so that they can configure it.

fix #572