feat(View): add support for point of views (#125)
The `PointOfView` object
========================

This object represents a point of view crafted for a specific camera
(perspective or orthographic), and may not apply to other cameras.

When using this object to setup the camera for which it was created,
the camera should fit the object in the viewport.

`View.goTo()`
=============

This method moves the camera to fit an object into the viewport.

- If the argument is an `Object3D`, then the `View` will compute the
camera parameters (distance to object, lookat target, etc) from the
bounding box of the object (and it's descendants).
- If the argument is a `PointOfView`, it will be used directly to setup
the camera.
- Finally, if the argument implements the `HasDefaultPointOfView` interface,
it will be used to retrieve the default point of view.

The `HasDefaultPointOfView` interface
=====================================

This interface allows object to provide a default `PointOfView` that
makes sense for the particular object.

It is implemented in `Entity3D` (that does something very similar to the
bounding box computation in the case of `Object3D`), and has specialized
implementations in `Map` (we use the map extent and elevation min/max to
compute the camera position), and for `Globe` (the camera is positioned
at the zenith of the (0°, 0°) geographic coordinate, and is looking at
the center of the globe).

BREAKING CHANGE: `Instance.focusObject()` is removed. Use `View.goTo()`
instead.`

```
instance.view.goTo(myObject);
```

fix #125