fix(Layer): use the source extent as fallback when constructing the LayerComposer
If the LayerComposer does not have any extent, then it cannot computes
the renderOrder (z-index) for images. This z-index is necessary when
rendering regions of the composer without knowing exactly which images
we need (for example, when generating a temporary texture until the real
one is available). In this scenario, all textures that overlap the
region as used, and we need a correct renderOrder to avoid putting
low resolution images on top of high resolution ones.