fix(Layer): apply default texture to node every time it is processed
If we are zooming out for example, the new node might already be known
to the layer, but it might not have any texture (or an obsolete texture).
We thus have to apply a default texture every time, and not the first
time this node is encountered by the Layer.