feat(entities): new entity OrientedImageCollection
This Entity can be used to display photos positionned and oriented images in the 3D space.
Visually, an image is made of 3 elements:
- a sphere at the location of the receptor
- a frustum
- a plane textured with the actual image

We want to be able to display potentially thousands of photos.
In an effort to reduce the number of draw call,
the spheres are an `InstancedMesh` and the frustum is an `InstancedMesh`.
The planes remain independent meshes since they each have their own texture.
This way, to display 1000 photos,
we only have 1002 draw calls (compared to 3000 if spheres and frustums were independent).

In order to easily manage those `InstancedMesh`, the photos are grouped in the PhotosEntity.