fix(COPCSource): fix bug where the entire file is queried
COPC format allows some nodes to be empty (they do not contain any point).
In this case, the spec here https://copc.io/#hierarchy-vlr states that
the node definition has to be {pointCount: 0, offset: 0, byteSize: 0}.

So when trying to load such a node,
Giro3D creates a range query as usual which goes from `offset` to `offset + bytesize - 1`,
which here is from `0` to `-1`. This results in a query `Range: bytes=0--1`,
which the browser interprets as `Range: bytes=0-`
which queries the whole file.