fix(TileFS.glsl): exclude shading from the shadow map code path (#579)
Typically, shadow maps render the same objects using different materials.

However, in the case of Map tiles (TileFS.glsl), the same shader code is
used for shadow mapping as well as regular rendering, to avoid having to
maintain two sets of complicated shaders.

Thus, the shader is running two versions of itself:

1. the normal version (specified by the COLOR_RENDER define), with all
  the features
2. the shadow mapping version (specified by the DISTANCE_RENDER and
  DEPTH_RENDER defines), that is only responsible for writing the depth
  buffer

However, some code paths were not properly excluded in the shadow mapping
version, leading to the same texture being read and written to, which is
an error in WebGL. On Linux, the error would be silently ignored. But on
MacOS (at least), the error would prevent the scene from being rendered.

fix #579