fix(RequestQueue): fix uncaught exception
If the signal provided by `RequestQueue.enqueue()` was aborted
before the `Task.execute()` is called,
then we end up calling `Promise.reject()` synchronously
which becomes an uncaught exception.

With this commit, in this situation,
we end up calling `Task.reject()` twice,
which should not be an issue
because rejecting or resolving a `Promise` several times is fine.