fix(RequestQueue): avoid uncaught error
If we call `onQueueAvailable()` here,
then it creates a form of recursion,
which can lead to a "maximum callstack exceeded" error
if there are many tasks for which `task.shouldExecute() === false`.

I think it is useless to call `this.onQueueAvailable` here
because we are in the loop that executes tasks anyways.