fix(PointsVS): make picking more robust
Now that will will start mixing coloring attributes,
simply checking for classification visibility is not enough
when deciding to ditch a point for gpu-picking.

The easiest fix is to compute color even when picking,
and then ditch the points that are transparent.
It is slightly more expensive but I don't think it's a problem.